+++ src/pages/TagsPage.js
@@ -83,13 +83,22 @@ export default function CategoriesPage(props) {
         console.log(response);
       })
       .catch((err) => {
-        if (err.response.status === 401) {
-          localStorage.clear();
-          history.push("/admin");
+        if (err.message !== undefined) {
+          if (err.message === "Network Error") {
+            alert.show("Network Error, please comeback later", {
+              type: "error",
+            });
+          }
+        }
+        if (err.response !== undefined) {
+          if (err.response.status === 401) {
+            localStorage.clear();
+            history.push("/admin");
 
-          alert.show("Your credentials expired, please login again", {
-            type: "error",
-          });
+            alert.show("Your credentials expired, please login again", {
+              type: "error",
+            });
+          }
         }
         console.log(err);
       });
@@ -115,17 +124,26 @@ export default function CategoriesPage(props) {
       })
       .catch((err) => {
         console.log(err);
-        if (err.response.status === 401) {
-          localStorage.clear();
-          history.push("/admin");
+        if (err.message !== undefined) {
+          if (err.message === "Network Error") {
+            alert.show("Network Error, please comeback later", {
+              type: "error",
+            });
+          }
+        }
+        if (err.response !== undefined) {
+          if (err.response.status === 401) {
+            localStorage.clear();
+            history.push("/admin");
 
-          alert.show("Your credentials expired, please login again", {
-            type: "error",
-          });
-        } else {
-          alert.show("Error: Check your internet connection", {
-            type: "error",
-          });
+            alert.show("Your credentials expired, please login again", {
+              type: "error",
+            });
+          } else {
+            alert.show("Error: Check your internet connection", {
+              type: "error",
+            });
+          }
         }
       });
   };
